requirements.txt
python-dotenv==1.0.1
langchain==0.1.13
langchain-openai==0.1.0
uvicorn==0.29.0
fastapi==0.110.0
beanie==1.25.0